task 4.21

+++ part4-exercises/BlogListAUTH-task/backend/controllers/blog.js
@@ -63,9 +63,45 @@ blogRouter.get('/:id', (request, response, next) => {
 })
 
 
+// blogRouter.delete('/:id', async (request, response) => {
+//     await Blog.findByIdAndDelete(request.params.id)
+//     response.status(204).end()
+// })
+
+// delete blog operation but blog can only be deleted by the user who created it
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id){
+        return response.status(401).json({
+            error: 'invalid token'
+        })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if(!user){
+        return response.status(404).json({
+            error: 'user not found'
+        })
+    }
+
+    if(!blog) {
+        return response.status(404).json({
+            error: 'blog not found'
+        })
+    }
+
+    if (blog.user.toString() === user.id.toString()){
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    }else{
+        return response.status(403).json({
+            error: 'blog cannot be deleted'
+        })
+    }
+
 })
 
 blogRouter.put('/:id', async (request, response) => {
